move to prevent viteDevServer reload
